WIP: Add routine to test circle-rectangle collision
I'm doing this because Raylib's built-in routine for collision detection seems to be erroneous. (Additionally, it's never performed well. In my 1,000 particle test runs, circle testing has been one of the more consistent bottlenecks despite the O(n^2) vs. O(n) time complexities [n = number of particles]. According to my inexpert reading of the profiler's reports, because of the time the CPU spends in converting between integers and floating-points as part of the implementation detail of the routine, which may be causing certain unknown effects that propagate outward.)

Or, more precisely, disk-rectangle collision (two-dimensional figures), which means that, if one figure is contained in the other, they are said to be colliding (intersecting).

(As opposed to the one-dimensional case where it would be considered not intersecting because no point of intersection is created on the respective curves.)

Remaining:
1. At least one affirmative case.
2. Coverage of every branch.
3. Rotated version of each test.

Later, I will compute the branch coverage for all tests.