Fix camera/rendering + apparent physical bias (#16)
* Introduce verlet.h (Verlet integration)

To investigate velocity bias issue.

Integrator does not seem to be cause of problem.

Why: Both Yoshida and Verlet have same problem.

* WIP: Add routine to test circle-rectangle collision

I'm doing this because Raylib's built-in routine for collision detection seems to be erroneous. (Additionally, it's never performed well. In my 1,000 particle test runs, circle testing has been one of the more consistent bottlenecks despite the O(n^2) vs. O(n) time complexities [n = number of particles]. According to my inexpert reading of the profiler's reports, because of the time the CPU spends in converting between integers and floating-points as part of the implementation detail of the routine, which may be causing certain unknown effects that propagate outward.)

Or, more precisely, disk-rectangle collision (two-dimensional figures), which means that, if one figure is contained in the other, they are said to be colliding (intersecting).

(As opposed to the one-dimensional case where it would be considered not intersecting because no point of intersection is created on the respective curves.)

Remaining:
1. At least one affirmative case.
2. Coverage of every branch.
3. Rotated version of each test.

Later, I will compute the branch coverage for all tests.

* Remove hypot from disk-rectangle intersection detection

The hypot calls were estimated to be O(n) where n is the number of particles in the previous implementation.

I will add in the tests soon when I find the time to do it.

* Revise disk and (area) rectangle testing routine

* Fix "flickering near boundaries" issue

The erroneous "CheckCollisionCircleRec" (Raylib) routine was replaced with a custom routine.

* circle.h: Remove unintended math function calls

The signbit calls were observed to be calling a system function that would extract the sign bit from the floating-point number.

* demo: Remove collision bias

Refresh the Halton disk at every frame to remove statistical bias when particles are partially colliding

* Fix last remaining camera issue

I ignored the screen offset, which is now taken into account.

* Table.h: Revert to Yoshida integration

Verlet integration was only used for exposition of a problem.